feat: rework comment util
